% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPQuery.R
\name{IPQuery}
\alias{IPQuery}
\title{Get IP information with IPQuery}
\usage{
IPQuery(ip = NULL, format = "text")
}
\arguments{
\item{ip}{IP address or "self" for the client's IP.  Can also be a vector of IPs.}

\item{format}{Desired output format ("text" or other supported formats). Defaults to "text".}
}
\value{
#' A list or a single element depending on the input `ip`.  If `ip` is a vector, a list of responses is returned. Otherwise, a single response is returned.  The response format depends on the `format` parameter.
}
\description{
This function queries the IPQuery API to get information about an IP address.
}
\details{
The function can be used to get information about the client's IP address or any other IP address.  The function supports multiple output formats including text, JSON, YAML, and XML.
}
