% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPQuery.R
\name{query}
\alias{query}
\title{Query a resource}
\usage{
query(i)
}
\arguments{
\item{i}{An object representing the resource to query.  The exact nature of this object depends on the `request()` and `req_perform()` functions used.}
}
\value{
#' A character string containing the response body.
}
\description{
This function queries a resource and returns the response body as a string.
}
\details{
Queries a resource and returns the response body as a string.
}
\keyword{internal}
